A merge to include the months old work on api routing, and the desktop app, as well as the status update for the project's recent revival.

+++ packages/profile.js
@@ -7,26 +7,34 @@ module.exports ={
     * @param {String} user The current username of the player.
     * @param {String} userID A numerical 14 digit value that is the reference to the player for the system. 
     */
-   create: function(user, userID){
-      copydir('../ref/player-template', `../profiles/${userID}`,{utimes: true, mode: true, cover: true},(err)=>{if(err){throw err;}});
-      setTimeout(function(){
-      var c = JSON.parse(fs.readFileSync(`../ref/player-template/profile.json`));
-      c.username = user;
-      c.ID = userID;
-      setTimeout(function(){
-         fs.writeFileSync(`../profiles/${userID}/profile.json`, JSON.stringify(c), (err) => {if(err) throw err});
-      }, 1000);
-      }, 1000);
-      
+   create: (user, userID) => {
+      if(!fs.exists(`../profiles/${userID}`)) {
+         copydir('../ref/player-template', `../profiles/${userID}`,{utimes: true, mode: true, cover: true},(err)=>{if(err){throw err;}});
+         setTimeout(() => {
+         var c = JSON.parse(fs.readFileSync(`../ref/player-template/profile.json`));
+         c.username = user;
+         c.ID = userID;
+         setTimeout(() => {
+            fs.writeFileSync(`../profiles/${userID}/profile.json`, JSON.stringify(c), (err) => {if(err) throw err});
+         }, 1000);
+         }, 1000);
+      }
    },
    /**
     * Retrieves the profile.json file from the referenced user's files.
     * @param {String} userID A numerical 14 digit value that is the reference to the player for the system.
     * @returns {JSON} The player's profile.json
     */
-   getprofile: function(userID){
+   getprofile: (userID) => {
       //console.log(fs.existsSync(`../profiles/${userID}/profile.json`));
       return JSON.parse(fs.readFileSync(`../profiles/${userID}/profile.json`));
+   },
+   /**
+    * Saves the given profile to its appropriate json file.
+    * @param {Profile} p The profile to be saved.
+    */
+   saveprofile: (p) => {
+      fs.writeFileSync(`../profiles/${p.ID}/profile.json`, JSON.stringify(p), (err) => {if(err) throw errr;});
    },
    /**
     * Functions for displaying attributes of the player's profile via text.
@@ -36,7 +44,7 @@ module.exports ={
        * Displays the basic, general information about the player.
        * @param {Profile} c The player's profile.
        */
-      basic: function(c){
+      basic: (c) => {
          var result =      `username:           ${c.username}\n`;
          result +=         `level-exp:           ${c.lvl}-${c.xp}\n`;
          result +=         `credits:            ${c.credits}\n`;
@@ -47,14 +55,14 @@ module.exports ={
        * Displays the coordinates and name of the location of the player's current position.
        * @param {Profile} c The player's profile.
        */
-      location: function(c){
+      location: (c) => {
          return            `location:           ${c.loc}, [${c.coords[0]},${c.coords[1]}]`;
       },
       /**
        * Displays the player's current resource amounts.
        * @param {Profile} c The player's profile.
        */
-      resources: function(c){
+      resources: (c) => {
          var result =      `personnel:          ${c.resources.personnel[0]}/${c.resources.personnel[1]}\n\n`;
 
          result +=         `helium 1:           ${c.resources.heliumI}\n`;
@@ -98,7 +106,7 @@ module.exports ={
        * Displays the player's current skill values.
        * @param {Profile} c The player's profile.
        */
-      skills: function(c){
+      skills: (c) => {
          var result =      `attacking:          ${c.skills.attacking}\n`;
          result +=         `defending:          ${c.skills.defending}\n\n`;
          
@@ -127,7 +135,7 @@ module.exports ={
     * @param {String} userID A numerical 14 digit value that is the reference to the player for the system. 
     * @param {String} choice The faction the player is choosing to join.
     */
-   chooseFaction: function(userID, choice){
+   chooseFaction: (userID, choice) => {
       var c = this.getprofile(userID);
       choice = choice.toLowerCase();
       switch(choice){
